refactor: extract audio logic into useVoiceVerification hook
