+++ apps/web/components/VoiceVerify.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useRef, useCallback } from "react";
+import { useVoiceVerification } from "../src/hooks/useVoiceVerification";
 
 type VerificationResult = {
     medicine_name_original: string;
@@ -49,102 +49,16 @@ const STATUS_CONFIG = {
 };
 
 export default function VoiceVerify() {
-    const [isRecording, setIsRecording] = useState(false);
-    const [isLoading, setIsLoading] = useState(false);
-    const [result, setResult] = useState<ApiResponse | null>(null);
-    const [error, setError] = useState<string | null>(null);
-    const [audioLevel, setAudioLevel] = useState(0);
-
-    const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-    const chunksRef = useRef<Blob[]>([]);
-    const animFrameRef = useRef<number | null>(null);
-    const analyserRef = useRef<AnalyserNode | null>(null);
-
-    const startRecording = useCallback(async () => {
-        setError(null);
-        setResult(null);
-
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-
-            // Visualize audio level
-            const audioCtx = new AudioContext();
-            const source = audioCtx.createMediaStreamSource(stream);
-            const analyser = audioCtx.createAnalyser();
-            analyser.fftSize = 256;
-            source.connect(analyser);
-            analyserRef.current = analyser;
-
-            const draw = () => {
-                const data = new Uint8Array(analyser.frequencyBinCount);
-                analyser.getByteFrequencyData(data);
-                const avg = data.reduce((a, b) => a + b, 0) / data.length;
-                setAudioLevel(avg / 128); // normalize 0–1
-                animFrameRef.current = requestAnimationFrame(draw);
-            };
-            draw();
-
-            // Start MediaRecorder
-            const recorder = new MediaRecorder(stream, { mimeType: "audio/webm" });
-            chunksRef.current = [];
-
-            recorder.ondataavailable = (e) => {
-                if (e.data.size > 0) chunksRef.current.push(e.data);
-            };
-
-            recorder.onstop = async () => {
-                stream.getTracks().forEach((t) => t.stop());
-                if (animFrameRef.current) cancelAnimationFrame(animFrameRef.current);
-                setAudioLevel(0);
-
-                const blob = new Blob(chunksRef.current, { type: "audio/webm" });
-                await sendAudioToApi(blob);
-            };
-
-            recorder.start();
-            mediaRecorderRef.current = recorder;
-            setIsRecording(true);
-        } catch {
-            setError("Microphone access denied. Please allow microphone access and try again.");
-        }
-    }, []);
-
-    const stopRecording = useCallback(() => {
-        if (mediaRecorderRef.current && isRecording) {
-            mediaRecorderRef.current.stop();
-            setIsRecording(false);
-        }
-    }, [isRecording]);
-
-    const sendAudioToApi = async (blob: Blob) => {
-        setIsLoading(true);
-        try {
-            const form = new FormData();
-            form.append("audio", blob, "recording.webm");
-
-            const res = await fetch("/api/medicine/verify-voice", {
-                method: "POST",
-                body: form,
-            });
-
-            const data: ApiResponse = await res.json();
-
-            if (!res.ok || !data.success) {
-                setError(data.error || "Verification failed. Please try again.");
-            } else {
-                setResult(data);
-            }
-        } catch {
-            setError("Network error. Please check your connection and try again.");
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    const reset = () => {
-        setResult(null);
-        setError(null);
-    };
+    const {
+        isRecording,
+        isLoading,
+        audioLevel,
+        result,
+        error,
+        startRecording,
+        stopRecording,
+        reset,
+    } = useVoiceVerification();
 
     const statusConfig = result ? STATUS_CONFIG[result.verification.status] : null;
 
@@ -192,7 +106,6 @@ export default function VoiceVerify() {
                               : "Tap to speak the medicine name"}
                     </p>
 
-                    {/* Supported languages hint */}
                     <p className="text-center text-xs text-gray-400">
                         Supports: Hindi • Tamil • Telugu • Kannada • Bengali • Malayalam + more
                     </p>
@@ -214,7 +127,6 @@ export default function VoiceVerify() {
                 <div
                     className={`rounded-2xl border-2 ${statusConfig.border} ${statusConfig.bg} space-y-4 p-5`}
                 >
-                    {/* Status Badge */}
                     <div className="flex items-center justify-between">
                         <span className={`text-lg font-bold ${statusConfig.text}`}>
                             {statusConfig.label}
@@ -227,7 +139,6 @@ export default function VoiceVerify() {
                         </span>
                     </div>
 
-                    {/* Medicine name in regional script */}
                     <div className="space-y-1">
                         <p className="text-xs tracking-wide text-gray-400 uppercase">
                             Medicine ({result.script} script)
@@ -244,7 +155,6 @@ export default function VoiceVerify() {
                         )}
                     </div>
 
-                    {/* Details */}
                     <div className="grid grid-cols-2 gap-3 text-sm">
                         <div>
                             <p className="text-xs text-gray-400">Manufacturer</p>
@@ -272,7 +182,6 @@ export default function VoiceVerify() {
                         </div>
                     </div>
 
-                    {/* Warnings */}
                     {result.verification.warnings.length > 0 && (
                         <div className="space-y-1 rounded-lg bg-white/60 p-3">
                             <p className="text-xs font-semibold text-gray-500 uppercase">
@@ -286,7 +195,6 @@ export default function VoiceVerify() {
                         </div>
                     )}
 
-                    {/* Try again */}
                     <button
                         onClick={reset}
                         className="w-full rounded-xl bg-gray-100 py-2 text-sm font-medium text-gray-700 transition hover:bg-gray-200"
@@ -296,7 +204,6 @@ export default function VoiceVerify() {
                 </div>
             )}
 
-            {/* Fallback text input */}
             {!result && !isRecording && (
                 <div className="text-center">
                     <p className="text-xs text-gray-400">
