+++ apps/web/package.json
@@ -19,7 +19,10 @@
         "@opentelemetry/auto-instrumentations-web": "^0.64.0",
         "@opentelemetry/context-zone": "^2.8.0",
         "@opentelemetry/exporter-trace-otlp-http": "^0.219.0",
+        "@opentelemetry/instrumentation": "^0.219.0",
         "@opentelemetry/instrumentation-fetch": "^0.219.0",
+        "@opentelemetry/resources": "^2.8.0",
+        "@opentelemetry/sdk-trace-base": "^2.8.0",
         "@opentelemetry/sdk-trace-web": "^2.8.0",
         "@supabase/auth-helpers-nextjs": "^0.15.0",
         "@supabase/ssr": "^0.12.0",
